UI: Move DropDown select to common component.
