+++ client/src/services/statusService.js
@@ -3,7 +3,6 @@ import axios from 'axios';
 class StatusService {
   static getStatus() {
     return axios({
-      method: 'get',
       url: process.env.REACT_APP_STATUS_SERVICE
     });
   }
